requirements.txt
aiofiles==24.1.0
aiosqlite==0.20.0
alabaster==1.0.0
asyncua==1.1.5
babel==2.16.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
docutils==0.21.2
evdev==1.7.1
freeopcua==0.90.6
idna==3.10
imagesize==1.4.1
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
opcua==0.98.13
opcua-client==0.8.4
opcua-widgets==0.6.1
packaging==24.2
pycparser==2.22
Pygments==2.19.1
pynput==1.7.7
pyOpenSSL==25.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
python-xlib==0.33
pytz==2024.2
requests==2.32.3
six==1.17.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tornado==6.4.2
typing_extensions==4.12.2
urllib3==2.3.0
rti.connext
